Refactored Path components structure
